Cosmetic changes to example scenes.
