+++ Assets/FancyScrollView/Examples/06_LoopTabBar.unity
@@ -127,16 +127,16 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 1141698468}
     m_Modifications:
-    - target: {fileID: 8026611260721418881, guid: de225ab998d364b3fb185d53340d8cd6,
+    - target: {fileID: 3194610462843345028, guid: de225ab998d364b3fb185d53340d8cd6,
         type: 3}
-      propertyPath: m_Name
-      value: Window2
+      propertyPath: m_Text
+      value: Window 2
       objectReference: {fileID: 0}
-    - target: {fileID: 8026611260721418881, guid: de225ab998d364b3fb185d53340d8cd6,
+    - target: {fileID: 4622212885601797420, guid: de225ab998d364b3fb185d53340d8cd6,
         type: 3}
-      propertyPath: m_IsActive
-      value: 0
-      objectReference: {fileID: 0}
+      propertyPath: graphicRaycaster
+      value: 
+      objectReference: {fileID: 1708390399}
     - target: {fileID: 6748594459359164585, guid: de225ab998d364b3fb185d53340d8cd6,
         type: 3}
       propertyPath: m_LocalPosition.x
@@ -247,15 +247,15 @@ PrefabInstance:
       propertyPath: graphicRaycaster
       value: 
       objectReference: {fileID: 1708390399}
-    - target: {fileID: 4622212885601797420, guid: de225ab998d364b3fb185d53340d8cd6,
+    - target: {fileID: 8026611260721418881, guid: de225ab998d364b3fb185d53340d8cd6,
         type: 3}
-      propertyPath: graphicRaycaster
-      value: 
-      objectReference: {fileID: 1708390399}
-    - target: {fileID: 3194610462843345028, guid: de225ab998d364b3fb185d53340d8cd6,
+      propertyPath: m_Name
+      value: Window2
+      objectReference: {fileID: 0}
+    - target: {fileID: 8026611260721418881, guid: de225ab998d364b3fb185d53340d8cd6,
         type: 3}
-      propertyPath: m_Text
-      value: Window 2
+      propertyPath: m_IsActive
+      value: 0
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: de225ab998d364b3fb185d53340d8cd6, type: 3}
@@ -288,6 +288,11 @@ PrefabInstance:
         type: 3}
       propertyPath: m_Name
       value: Dropdown
+      objectReference: {fileID: 0}
+    - target: {fileID: 1747360886316832958, guid: a5f73e63b9cf5418ca91edef13ad6686,
+        type: 3}
+      propertyPath: defaultScene
+      value: 5
       objectReference: {fileID: 0}
     - target: {fileID: 1747360886316832959, guid: a5f73e63b9cf5418ca91edef13ad6686,
         type: 3}
@@ -393,11 +398,6 @@ PrefabInstance:
         type: 3}
       propertyPath: m_Pivot.y
       value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 1747360886316832958, guid: a5f73e63b9cf5418ca91edef13ad6686,
-        type: 3}
-      propertyPath: defaultScene
-      value: 5
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: a5f73e63b9cf5418ca91edef13ad6686, type: 3}
@@ -733,16 +733,16 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 1141698468}
     m_Modifications:
-    - target: {fileID: 8026611260721418881, guid: de225ab998d364b3fb185d53340d8cd6,
+    - target: {fileID: 3194610462843345028, guid: de225ab998d364b3fb185d53340d8cd6,
         type: 3}
-      propertyPath: m_Name
-      value: Window0
+      propertyPath: m_Text
+      value: Window 0
       objectReference: {fileID: 0}
-    - target: {fileID: 8026611260721418881, guid: de225ab998d364b3fb185d53340d8cd6,
+    - target: {fileID: 4622212885601797420, guid: de225ab998d364b3fb185d53340d8cd6,
         type: 3}
-      propertyPath: m_IsActive
-      value: 0
-      objectReference: {fileID: 0}
+      propertyPath: graphicRaycaster
+      value: 
+      objectReference: {fileID: 1708390399}
     - target: {fileID: 6748594459359164585, guid: de225ab998d364b3fb185d53340d8cd6,
         type: 3}
       propertyPath: m_LocalPosition.x
@@ -853,15 +853,15 @@ PrefabInstance:
       propertyPath: graphicRaycaster
       value: 
       objectReference: {fileID: 1708390399}
-    - target: {fileID: 4622212885601797420, guid: de225ab998d364b3fb185d53340d8cd6,
+    - target: {fileID: 8026611260721418881, guid: de225ab998d364b3fb185d53340d8cd6,
         type: 3}
-      propertyPath: graphicRaycaster
-      value: 
-      objectReference: {fileID: 1708390399}
-    - target: {fileID: 3194610462843345028, guid: de225ab998d364b3fb185d53340d8cd6,
+      propertyPath: m_Name
+      value: Window0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8026611260721418881, guid: de225ab998d364b3fb185d53340d8cd6,
         type: 3}
-      propertyPath: m_Text
-      value: Window 0
+      propertyPath: m_IsActive
+      value: 0
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: de225ab998d364b3fb185d53340d8cd6, type: 3}
@@ -995,8 +995,8 @@ GameObject:
   - component: {fileID: 2133201406}
   - component: {fileID: 2133201411}
   - component: {fileID: 2133201410}
-  - component: {fileID: 2133201409}
   - component: {fileID: 2133201407}
+  - component: {fileID: 2133201409}
   m_Layer: 5
   m_Name: Tabs
   m_TagString: Untagged
@@ -1112,16 +1112,16 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 1141698468}
     m_Modifications:
-    - target: {fileID: 8026611260721418881, guid: de225ab998d364b3fb185d53340d8cd6,
+    - target: {fileID: 3194610462843345028, guid: de225ab998d364b3fb185d53340d8cd6,
         type: 3}
-      propertyPath: m_Name
-      value: Window1
+      propertyPath: m_Text
+      value: Window 1
       objectReference: {fileID: 0}
-    - target: {fileID: 8026611260721418881, guid: de225ab998d364b3fb185d53340d8cd6,
+    - target: {fileID: 4622212885601797420, guid: de225ab998d364b3fb185d53340d8cd6,
         type: 3}
-      propertyPath: m_IsActive
-      value: 0
-      objectReference: {fileID: 0}
+      propertyPath: graphicRaycaster
+      value: 
+      objectReference: {fileID: 1708390399}
     - target: {fileID: 6748594459359164585, guid: de225ab998d364b3fb185d53340d8cd6,
         type: 3}
       propertyPath: m_LocalPosition.x
@@ -1232,15 +1232,15 @@ PrefabInstance:
       propertyPath: graphicRaycaster
       value: 
       objectReference: {fileID: 1708390399}
-    - target: {fileID: 4622212885601797420, guid: de225ab998d364b3fb185d53340d8cd6,
+    - target: {fileID: 8026611260721418881, guid: de225ab998d364b3fb185d53340d8cd6,
         type: 3}
-      propertyPath: graphicRaycaster
-      value: 
-      objectReference: {fileID: 1708390399}
-    - target: {fileID: 3194610462843345028, guid: de225ab998d364b3fb185d53340d8cd6,
+      propertyPath: m_Name
+      value: Window1
+      objectReference: {fileID: 0}
+    - target: {fileID: 8026611260721418881, guid: de225ab998d364b3fb185d53340d8cd6,
         type: 3}
-      propertyPath: m_Text
-      value: Window 1
+      propertyPath: m_IsActive
+      value: 0
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: de225ab998d364b3fb185d53340d8cd6, type: 3}
